Conflicts:
	.gitignore
	Applications/Present/PresentApp.cpp
	Applications/ViewBasic/Main.cpp
	CornerStone.pri
	CornerStone.pro
	Examples/Widget3D/PropellerWidget.hpp
	Fluffy/Grammar.cpp
	Fluffy/Grammar.hpp
	MultiBrowser/BrowserWidget.cpp
	MultiTouch/BGCalibrator.cpp
	MultiTouch/MultiTouch.cpp
	MultiTouch/MultiTouch.pro
	MultiTouch/TrackingSource.hpp
	MultiWidgets/BigImageMovie.cpp
	MultiWidgets/BigImageMovie.hpp
	MultiWidgets/Calibration.cpp
	MultiWidgets/GrabManager.cpp
	MultiWidgets/ImageWidget.cpp
	MultiWidgets/Keyboard.cpp
	MultiWidgets/Mime.cpp
	MultiWidgets/MultiWidgets.pro
	MultiWidgets/Plugins.cpp
	MultiWidgets/Plugins.hpp
	MultiWidgets/RippleOverlay.cpp
	MultiWidgets/SDLApplication.cpp
	MultiWidgets/SDLApplication.hpp
	MultiWidgets/TextBox.cpp
	MultiWidgets/VideoWidget.cpp
	MultiWidgets/Widget.cpp
	MultiWidgets/Widget3D.cpp
	Plugins/Book3D/BookWidget.cpp
	Plugins/Effects/Effects.cpp
	Plugins/Effects/Effects.pro
	Plugins/Effects/SpriteRenderer.cpp
	Plugins/FingerPhysics2/Particles.cpp
	Plugins/HandMenu/HandMenu.cpp
	Plugins/Plugins.pri
	Scripting/generator/exclude_files
	ThreadedRendering/RenderThread.cpp
	multitude/Applications/MoviePlayer/VideoWindow.cpp
	multitude/Luminous/CPUMipmaps.cpp
	multitude/Luminous/CPUMipmaps.hpp
	multitude/Luminous/Error.cpp
	multitude/Luminous/GLResources.hpp
	multitude/Luminous/GLSLProgramObject.cpp
	multitude/Luminous/GLSLProgramObject.hpp
	multitude/Luminous/GPUMipmaps.cpp
	multitude/Luminous/GPUMipmaps.hpp
	multitude/Luminous/Image.cpp
	multitude/Luminous/Image.hpp
	multitude/Luminous/ImageCodecPNG.cpp
	multitude/Luminous/ImageCodecSVG.cpp
	multitude/Luminous/ImageCodecTGA.cpp
	multitude/Luminous/Luminous.cpp
	multitude/Luminous/Luminous.pro
	multitude/Luminous/MatrixStep.hpp
	multitude/Luminous/MultiHead.cpp
	multitude/Luminous/MultiHead.hpp
	multitude/Luminous/PixelFormat.cpp
	multitude/Luminous/PixelFormat.hpp
	multitude/Luminous/RenderContext.cpp
	multitude/Luminous/RenderContext.hpp
	multitude/Luminous/Shader.cpp
	multitude/Luminous/Shader.hpp
	multitude/Luminous/Texture.cpp
	multitude/Luminous/Texture.hpp
	multitude/Luminous/Utils.cpp
	multitude/Luminous/VertexBuffer.hpp
	multitude/Luminous/VertexBufferImpl.hpp
	multitude/Nimble/Math.hpp
	multitude/Nimble/Matrix3.hpp
	multitude/Nimble/Matrix4.hpp
	multitude/Nimble/Matrix4Impl.hpp
	multitude/Poetic/CPUFont.cpp
	multitude/Poetic/CPUWrapperFont.cpp
	multitude/Poetic/GPUManagedFont.cpp
	multitude/Poetic/GPUManagedFont.hpp
	multitude/Poetic/GPUTextureFont.cpp
	multitude/Radiant/CameraDriver.cpp
	multitude/Radiant/FixedStr.hpp
	multitude/Radiant/FixedStrImpl.hpp
	multitude/Radiant/Mutex.hpp
	multitude/Radiant/Platform.hpp
	multitude/Radiant/PlatformUtilsOSX.cpp
	multitude/Radiant/Radiant.pro
	multitude/Radiant/TimeStamp.hpp
	multitude/VideoDisplay/ShowGL.hpp
	multitude/library.pri
	multitude/multitude.pri

+++ multitude/Poetic/GlyphContainer.cpp
@@ -37,7 +37,7 @@ namespace Poetic
 
   void GlyphContainer::add(Glyph * glyph, unsigned int charCode)
   {
-    m_charmap->insertIndex(charCode, m_glyphs.size());
+    m_charmap->insertIndex(charCode, static_cast<unsigned int> (m_glyphs.size()));
     m_glyphs.push_back(glyph);
   }
